requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyjwt
passlib[bcrypt]
email-validator
sqlalchemy
pymysql
